requirements.txt
streamlit>=1.30.0
python-pptx>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
